Minor cleanup + some written thoughts

+++ src/parser/RuleState.cc
@@ -568,7 +568,7 @@ void OrState::handleRootResult(SharedRuleState p_clone)
 // =====================================================================================================================
 void OrState::saveToNewSlot(SharedRuleState p_clonedState)
 {
-    unsigned size = m_slots.size();
+    size_t size = m_slots.size();
     m_slots.resize(size + 1);
     static_cast<AndOrCommon*>(p_clonedState.get())->updateParentSlot(this, this, size);
     createTypedSlot(SlotType::One, size, p_clonedState);
